Refactoring ROV to have source files and spliting it down a little.
